+++ BitbucketServerAutomation/Functions/Add-PSTypeName.ps1
@@ -12,27 +12,33 @@
 
 function Add-PSTypeName
 {
+    [CmdletBinding()]
     param(
         [Parameter(Mandatory=$true,ValueFromPipeline=$true)]
         $InputObject,
 
+        [Parameter(Mandatory=$true,ParameterSetName='RepositoryInfo')]
         [Switch]
         $RepositoryInfo,
 
+        [Parameter(Mandatory=$true,ParameterSetName='ProjectInfo')]
         [Switch]
-        $ProjectInfo
+        $ProjectInfo,
+
+        [Parameter(Mandatory=$true,ParameterSetName='CommmitBuildStatusInfo')]
+        [Switch]
+        $CommitBuildStatusInfo
     )
 
     process
     {
-        if( $RepositoryInfo )
-        {
-            $InputObject.pstypenames.Add( 'Atlassian.Bitbucket.Server.RepositoryInfo' )
-        }
+        Set-StrictMode -Version 'Latest'
+
+        $typeName = 'Atlassian.Bitbucket.Server.{0}' -f $PSCmdlet.ParameterSetName
+        $InputObject.pstypenames.Add( $typeName )
 
         if( $ProjectInfo )
         {
-            $InputObject.pstypenames.Add( 'Atlassian.Bitbucket.Server.ProjectInfo' )
             if( -not ($InputObject | Get-Member -Name 'description') )
             {
                 $InputObject | Add-Member -MemberType NoteProperty -Name 'description' -Value ''